Splitting the experiecne and excluseve serive section from main page
